analyzers_manager(scripts): VirusTotalv3AnalyzerMixin class
